+++ src/events/RegisterVariablesEvent.php
@@ -1,6 +1,10 @@
 <?php
 namespace verbb\formie\events;
 
+use verbb\formie\compatibility\variables\VariableSourceCompatibility;
+use verbb\formie\variables\VariableSource;
+use verbb\formie\variables\VariableSourceInterface;
+
 use yii\base\Event;
 
 class RegisterVariablesEvent extends Event
@@ -8,6 +12,6 @@ class RegisterVariablesEvent extends Event
     // Properties
     // =========================================================================
 
-    public array $variables = [];
-    
+    public array $sources = [];
+
 }